requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alpha_vantage==3.0.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
dateparser==1.2.0
Django==5.1.3
django-allauth==65.3.0
django-daraja==1.3.0
django-widget-tweaks==1.5.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
plotly==5.24.1
propcache==0.2.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
tenacity==9.0.0
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
websockets==14.1
yarl==1.18.3
